LPT-2370 - Move ALERT_PATTERN to a shared constants class
Move ALERT_PATTERN out of CourtListPublishAndPDFGenerationTask and into
AppConstant, since it's an alerting marker shared by both the standard and
SJP publish flows - keeping it on a single task class was misleading. All
six consumers (CaTHService, CourtListPdfHelper, SjpCourtListPublishService,
SjpTaskTriggerService, CourtListPublishAndPDFGenerationTask, SjpPublishTask)
now import it from AppConstant instead.
