+++ src/main/java/uk/gov/hmcts/cp/config/AppConstant.java
@@ -4,4 +4,7 @@ public class AppConstant {
     private AppConstant(){}
 
     public static final String CJSCPPUID = "CJSCPPUID";
+
+    /** Log marker for alerting on publish failures; used across both the standard and SJP flows. */
+    public static final String ALERT_PATTERN = "PUBLISHING_FAILED";
 }
